+++ blinkreceipt-demo/BlinkReceiptSurveys/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.microblink.blinkreceiptsurveys">
+    package="com.blinkreceipt.development">
 
   <application
       android:name=".BlinkSurveysApplication"
@@ -19,8 +19,8 @@
     </activity>
 
     <meta-data
-        android:name="com.microblink.ProductIntelligence"
-        android:value="[PRODUCT INTEL KEY]" />
+        android:name="com.microblink.LicenseKey"
+        android:value="sRwAAAAYY29tLmJsaW5rcmVjZWlwdC5zdXJ2ZXlzh1Abwgh8le/dWIol0Xw6wvpMplT4/2UcI8gnypHpg/eWVnEKvWBeso66PqE/78CB/7UZ5WP/I7TI/xNva3Mt4cwiorD9JlEgVk/4DJTGx8k2/cGm3xHtwGs2NpUzjxbSPtKYLVkrk/pEDl2j6chR6jfOpyWnCu9cOAp/mMAfEQbStit/Pq8=" />
 
     <meta-data
         android:name="com.microblink.ClientUserId"
