+++ blinkreceipt-demo/BlinkReceiptSurveys/src/main/res/layout/activity_main.xml
@@ -4,8 +4,16 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
+    tools:context="com.blinkreceipt.development.MainActivity"
     >
+
+  <Button
+      android:id="@+id/load_survey_btn"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      app:layout_constraintStart_toStartOf="parent"
+      app:layout_constraintTop_toTopOf="parent"
+      />
 
   <FrameLayout
       android:id="@+id/survey_id_container"
