--- blinkreceipt-demo/blink-receipt-core/build.gradle
@@ -1,2 +1,0 @@
-configurations.maybeCreate("default")
-artifacts.add("default", file('LibBlinkReceiptCore.aar'))
