--- blinkreceipt-demo/blink-receipt-surveys/build.gradle
@@ -1,2 +1,0 @@
-configurations.maybeCreate("default")
-artifacts.add("default", file('LibBlinkReceiptSurveys.aar'))
